test: split PDF visual native rendition coverage
